[BOT] Reorganizing Codebase Without Disrupting Functionality
- [+] refactor(chat.tsx): extract usePinApp and useDebouncedEffect functions to separate files
